Extensions & imports for Matrix; move Vec to Matrix.
